requirements.txt
pandas>=2.2.0
numpy>=1.26.3
plotly>=5.18.0
anthropic>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pytest>=7.4.4
pytest-cov>=4.1.0

